# docs/swagger3.0.yaml
openapi: 3.0.0
info:
  version: 1.0.16
  title: The Student Food forum
  description: >
    A forum API where accounts can post recipts about food and comment to them.  

    The application uses bearer token to authenticate requests.
    This token is created by requests to the /session router.

    All data is bound to the account and if ever deleted all other data will be
    cascaded.
    Comments are bound to the recipe and will also cascade if recipe is deleted.
  contact:
    name: "Filip Andersson, Marcus Gullstrand"
...
servers:
  - url: thesff.us-east-1.elasticbeanstalk.com
paths:
  /accounts:
    get:
      tags:
        - accounts
      summary: Get accounts.
      description: ""
      parameters:
        - in: query
          name: search
          description: Query string that will filter search result on account nickname.
          schema:
            type: string
        - in: query
          name: limit
          description: How many accounts will be returned.
          schema:
            type: integer
            default: 25
        - in: query
          name: offset
          description: How many accounts that will be skipped.
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/accounts"
        "500":
          description: Internal server error!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
    post:
      tags:
        - accounts
      summary: Create a new account.
      responses:
        "201":
          description: Account created.
          headers:
            Location:
              schema:
                type: string
        "400":
          description: Missing required input parameter in body!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "409":
          description: Email already taken!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/postAccount"
              properties:
                nickname:
                  type: string
  "/accounts/{accountId}":
    patch:
      tags:
        - accounts
      summary: Edit given account.
      description: ""
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: The id of the account.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Missing parameter!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "401":
          description: Unauthorized
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                nickname:
                  description: >-
                    Edit the account nickname. If left blank it will not be
                    edited.
                  type: string
                password:
                  description: >-
                    Edit the account password. If left blank it will not be
                    edited.
                  type: string
                image:
                  description: >-
                    Add/edit the account image. If left blank it will not be
                    added/edited.
                  type: string
                  format: binary
    delete:
      tags:
        - accounts
      summary: Delete given account.
      description: >-
        Deleting a user will delete all of the users uploaded content (account,
        recipes, comments and votes).
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: Is of the account
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account succsessfully deleted
        "401":
          description: Unauthorized
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "404":
          description: Account was not found!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
...
  /comments:
    get:
      tags:
        - comments
      summary: Get all comments.
      description: ""
      parameters:
        - in: query
          name: recipesId
          required: false
          description: Get all comments for a specific recipesId
          schema:
            type: string
        - in: query
          name: limit
          description: How many comments will be returned.
          schema:
            type: integer
            default: 25
        - in: query
          name: offset
          description: How many comments that will be skipped.
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/getComments"
        "400":
          description: Missing parameter!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
    post:
      tags:
        - comments
      summary: Create a new comment.
      description: ""
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Comment successfully created!
        "400":
          description: Missing parameter!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "401":
          description: Unauthorized
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/postComments"
  "/comments/{commentId}":
    patch:
      tags:
        - comments
      summary: Edit a given account.
      description: ""
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          description: comment Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully changed!
        "400":
          description: Missing required parameter!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "401":
          description: Unauthorized
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
        description: The content of comment
    delete:
      tags:
        - comments
      summary: Delete a given account.
      description: ""
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          description: comment id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Missing required parameter!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "401":
          description: Unauthorized
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
...
components:
  schemas:
    error:
      type: object
      properties:
        error:
          type: string
    postAccount:
      type: object
      properties:
        email:
          type: string
          format: email
          example: foo@gmail.com
        password:
          type: string
          format: password
      required:
        - password
        - email
...
    getComments:
      type: object
      properties:
        accountId:
          type: string
        recipeId:
          type: string
        content:
          type: string
        commentId:
          type: string
        updatedAt:
          type: string
    postComments:
      type: object
      properties:
        content:
          type: string
        recipeId:
          type: string
      required:
        - accountId
        - recipeId
    accounts:
      type: object
      properties:
        id:
          type: string
        nickname:
          type: string
        imageId:
          type: string
...